Refactored NewsViewController
