+++ Client/Posts List/NewsViewController.swift
@@ -15,11 +15,8 @@ import Kingfisher
 import HNScraper
 import Loaf
 
-class NewsViewController : UIViewController {
+class NewsViewController : BaseTableViewController {
     public var hackerNewsService: HackerNewsService?
-    
-    @IBOutlet weak var tableView: UITableView!
-    private var refreshControl: UIRefreshControl!
     
     var posts: [HNPost] = [HNPost]()
     var postType: HNScraper.PostListPageName! = .news
@@ -29,13 +26,8 @@ class NewsViewController : UIViewController {
     
     override func viewDidLoad() {
         super.viewDidLoad()
-        
         registerForPreviewing(with: self, sourceView: tableView)
-
-        refreshControl = UIRefreshControl()
-        refreshControl.addTarget(self, action: #selector(NewsViewController.loadPosts), for: UIControl.Event.valueChanged)
-        tableView.refreshControl = refreshControl
-        
+        self.tableView.refreshControl?.addTarget(self, action: #selector(loadPosts), for: UIControl.Event.valueChanged)
         setupTheming()
         loadPosts()
     }
@@ -48,30 +40,20 @@ class NewsViewController : UIViewController {
         }
     }
     
-    override func viewWillTransition(to size: CGSize, with coordinator: UIViewControllerTransitionCoordinator) {
-        DispatchQueue.global().async(execute: {
-            DispatchQueue.main.sync {
-                self.viewDidRotate()
+    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
+        if segue.identifier == "ShowComments" {
+            if let indexPath = tableView.indexPathForSelectedRow,
+                let segueNavigationController = segue.destination as? UINavigationController,
+                let commentsViewController = segueNavigationController.topViewController as? CommentsViewController {
+                let post = posts[indexPath.row]
+                commentsViewController.post = post
             }
-        })
-    }
-    
-    public func viewDidRotate() {
-        guard let tableView = self.tableView, let indexPaths = tableView.indexPathsForVisibleRows else { return }
-        self.tableView.beginUpdates()
-        self.tableView.reloadRows(at: indexPaths, with: .automatic)
-        self.tableView.endUpdates()
-    }
-
-    func getSafariViewController(_ url: URL) -> SFSafariViewController {
-        let safariViewController = ThemedSafariViewController(url: url)
-        safariViewController.previewActionItemsDelegate = self
-        return safariViewController
+        }
     }
 }
 
 extension NewsViewController { // post fetching
-    @objc func loadPosts() {
+    @objc public func loadPosts() {
         hackerNewsService?.getPosts(of: self.postType).map { (posts, nextPageIdentifier) in
             self.posts = posts ?? [HNPost]()
             self.nextPageIdentifier = nextPageIdentifier
@@ -100,25 +82,14 @@ extension NewsViewController { // post fetching
             Loaf("Error connecting to Hacker News", state: .error, sender: self).show()
         }
     }
-    
-    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
-        if segue.identifier == "ShowComments" {
-            if let indexPath = tableView.indexPathForSelectedRow,
-            let segueNavigationController = segue.destination as? UINavigationController,
-            let commentsViewController = segueNavigationController.topViewController as? CommentsViewController {
-                let post = posts[indexPath.row]
-                commentsViewController.post = post
-            }
-        }
-    }
 }
 
-extension NewsViewController: UITableViewDataSource {
-    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
+extension NewsViewController {
+    override func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
         return posts.count
     }
     
-    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
+    override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
         let cell = tableView.dequeueReusableCell(withIdentifier: "PostCell", for: indexPath) as! PostCell
         cell.delegate = self
         cell.clearImage()
@@ -132,8 +103,8 @@ extension NewsViewController: UITableViewDataSource {
     }
 }
 
-extension NewsViewController: UITableViewDelegate {
-    func tableView(_ tableView: UITableView, willDisplay cell: UITableViewCell, forRowAt indexPath: IndexPath) {
+extension NewsViewController {
+    override func tableView(_ tableView: UITableView, willDisplay cell: UITableViewCell, forRowAt indexPath: IndexPath) {
         if indexPath.row == posts.count - 5 {
             loadMorePosts()
         }
@@ -142,10 +113,10 @@ extension NewsViewController: UITableViewDelegate {
 
 extension NewsViewController: Themed {
     func applyTheme(_ theme: AppTheme) {
-        view.backgroundColor = theme.backgroundColor
-        tableView.backgroundColor = theme.backgroundColor
-        tableView.separatorColor = theme.separatorColor
-        refreshControl.tintColor = theme.appTintColor
+        self.view.backgroundColor = theme.backgroundColor
+        self.tableView.backgroundColor = theme.backgroundColor
+        self.tableView.separatorColor = theme.separatorColor
+        self.tableView.refreshControl?.tintColor = theme.appTintColor
     }
 }
 
@@ -177,9 +148,15 @@ extension NewsViewController: UIViewControllerPreviewingDelegate, SFSafariViewCo
         }
         return [viewCommentsPreviewAction]
     }
+    
+    func getSafariViewController(_ url: URL) -> SFSafariViewController {
+        let safariViewController = ThemedSafariViewController(url: url)
+        safariViewController.previewActionItemsDelegate = self
+        return safariViewController
+    }
 }
 
-extension NewsViewController: PostTitleViewDelegate {
+extension NewsViewController: PostTitleViewDelegate, PostCellDelegate {
     func didPressLinkButton(_ post: HNPost) {
         guard verifyLink(post.url), let url = post.url else { return }
         self.navigationController?.present(getSafariViewController(url), animated: true, completion: nil)
@@ -189,9 +166,7 @@ extension NewsViewController: PostTitleViewDelegate {
         guard let url = url else { return false }
         return UIApplication.shared.canOpenURL(url)
     }
-}
-
-extension NewsViewController: PostCellDelegate {
+    
     func didTapThumbnail(_ sender: Any) {
         guard let tapGestureRecognizer = sender as? UITapGestureRecognizer else { return }
         let point = tapGestureRecognizer.location(in: tableView)
